+++ example/src/main/kotlin/com/animals/Animal.kt
@@ -1,0 +1,7 @@
+package com.animals
+
+public interface Animal {
+    val name: String
+    val isExtinct: Boolean
+    fun makeSound(): String
+}