+++ example/src/main/kotlin/com/animals/pets/Dog.kt
@@ -1,0 +1,8 @@
+package com.animals.pets
+
+internal class Dog : Pet() {
+    override val name: String = "DOG"
+    override fun makeSound(): String {
+        return "Wuff"
+    }
+}