+++ example/src/main/kotlin/com/firstexample/first/CompanionExamples.kt
@@ -1,0 +1,13 @@
+package com.firstexample.first
+
+interface Example {
+    val number: Int
+}
+
+class CompanionExamples {
+
+    companion object : Example {
+        override val number: Int
+            get() = 0
+    }
+}