+++ pumlgenerator/src/main/kotlin/uml/element/EnumElement.kt
@@ -6,6 +6,8 @@ import com.google.devtools.ksp.symbol.ClassKind
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
+import filterFunctionsByOptions
+import filterPropertiesByOptions
 import isValid
 import uml.DiagramElement
 import uml.ElementKind
@@ -36,18 +38,30 @@ $functionsString
 """
     }
 
-    class Builder(override val clazz: KSClassDeclaration, override val options: Options? = null, val logger: KSPLogger?) : DiagramElement.Builder<EnumElement> {
+    class Builder(override val clazz: KSClassDeclaration, override val options: Options, val logger: KSPLogger?) : DiagramElement.Builder<EnumElement> {
         val companionObject = clazz.declarations.filter { it is KSClassDeclaration && it.isCompanionObject }.map { it as? KSClassDeclaration }.firstOrNull()
 
-        override val properties: MutableList<KSPropertyDeclaration> = mutableListOf<KSPropertyDeclaration>().apply {
-            addAll((companionObject?.getAllProperties() ?: emptySequence()).filter { options?.isValid(it, logger) == true })
-            addAll(clazz.getAllProperties().filter { options?.isValid(it, logger) == true })
-        }
+        override val properties: MutableList<KSPropertyDeclaration> = buildList {
+            val validCompanionObjectProperties = companionObject?.getAllProperties()?.filterPropertiesByOptions(clazz, options, logger) ?: emptySequence()
+            addAll(validCompanionObjectProperties)
 
-        override val functions: MutableList<KSFunctionDeclaration> = mutableListOf<KSFunctionDeclaration>().apply {
-            addAll((companionObject?.getAllFunctions() ?: emptySequence()).filter { options?.isValid(it, logger) == true })
-            addAll(clazz.getAllFunctions().filter { options?.isValid(it, logger) == true })
-        }
+            val validProperties = clazz.getAllProperties().filterPropertiesByOptions(clazz, options, logger)
+            addAll(validProperties)
+
+            val otherProperties = clazz.declarations.filterIsInstance<KSPropertyDeclaration>().filterNot { it in clazz.getAllProperties() }.filterPropertiesByOptions(clazz, options, logger)
+            addAll(otherProperties)
+        }.toMutableList()
+
+        override val functions: MutableList<KSFunctionDeclaration> = buildList {
+            val validCompanionObjectFunctions = companionObject?.getAllFunctions()?.filterFunctionsByOptions(clazz, options, logger) ?: emptySequence()
+            addAll(validCompanionObjectFunctions)
+
+            val validFunctions = clazz.getAllFunctions().filterFunctionsByOptions(clazz, options, logger)
+            addAll(validFunctions)
+
+            val otherFunctions = clazz.declarations.filterIsInstance<KSFunctionDeclaration>().filterNot { it in clazz.getAllFunctions() }.filterFunctionsByOptions(clazz, options, logger)
+            addAll(otherFunctions)
+        }.toMutableList()
 
         override fun build(): EnumElement? {
             return if (options.isValid(clazz, logger)) {