+++ pumlgenerator/src/main/kotlin/uml/element/Field.kt
@@ -1,7 +1,6 @@
 package uml.element
 
 import Options
-import com.google.devtools.ksp.closestClassDeclaration
 import com.google.devtools.ksp.getVisibility
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
@@ -22,7 +21,8 @@ data class Field(val originalKSProperty: KSPropertyDeclaration, val showVisibili
     val modifiers = mutableListOf<String>().apply {
         val isParentCompanion = (originalKSProperty.parent as? KSClassDeclaration)?.isCompanionObject == true
         val isExtensionReceiverCompanion = (originalKSProperty.extensionReceiver?.resolve()?.declaration as? KSClassDeclaration)?.isCompanionObject == true
-        val isCompanionFunction = isParentCompanion || isExtensionReceiverCompanion
+        val isJVMStatic = originalKSProperty.modifiers.contains(Modifier.JAVA_STATIC)
+        val isCompanionFunction = isParentCompanion || isExtensionReceiverCompanion || isJVMStatic
         if (isCompanionFunction) {
             add("{static}")
         }