+++ pumlgenerator/src/test/kotlin/CompilationTest.kt
@@ -1,12 +1,10 @@
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import com.tschuchort.compiletesting.kspProcessorOptions
-import com.tschuchort.compiletesting.sourcesGeneratedBySymbolProcessor
 import com.tschuchort.compiletesting.symbolProcessorProviders
 import com.tschuchort.compiletesting.useKsp2
 import org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi
 import org.jetbrains.kotlin.incremental.createDirectory
-import org.jetbrains.kotlin.ir.types.IdSignatureValues.result
 import java.io.File
 import kotlin.test.assertFalse
 
@@ -31,5 +29,5 @@ open class CompilationTest {
 }
 
 internal fun assertContainsNot(charSequence: CharSequence, other: CharSequence, ignoreCase: Boolean = false, message: String? = null) {
-    assertFalse { charSequence.contains(other, ignoreCase = ignoreCase) }
+    assertFalse(charSequence.contains(other, ignoreCase = ignoreCase),  "\"Expected the char sequence to not contain the substring.\nCharSequence <$charSequence>, substring <$other>, ignoreCase <$ignoreCase>.\"")
 }